Final state before submission
This is the state of the project before its submission to the CS 61A
recursive art contest. It was cleaned up for submission. Here there is
a commented out section dealing with 3D representations—This was going
to be the next thing implemented, but I determined that implementing it
would take me over the token limit (at the time I was under the
impression that parentheses counted as tokens for the contest’s
limits). Several examples are included in system.scm. Also included is
scheme_primitives.py. In this file are the additional mappings I added
to the vanilla scheme interpreter so that it would be able to render
graphics and do necessary calculations. As of now, example operations
are in system.scm and graphic-ops.scm only includes stack elements for
branching. The idea for the final iteration was for the latter file to
be a module containing all operations, so this should be one fix to
come.